Will the refactoring works well ?

+++ CLI/AuxDataEntry.cs
@@ -3,8 +3,8 @@
 public struct AuxDataEntry
 {
     public ulong Time;
-    public ulong Pid;
-    public ulong Tid;
-    public ulong Cpu;
+    public uint Pid;
+    public uint Tid;
+    public uint Cpu;
     public ulong Flags;
 }
